### 2024-11-23 #7521 @danielskatz
👋 @jsdodge - thanks for your submission. You may have seen that our DOI checker has found a citation in your .bib file without a DOI. Given that you don't use this in the paper, you can either add the DOI or remove it from the .bib file.

Once you do either one, please use the command `@editorialbot check references` to check again. editorialbot commands need to be the first entry in a new comment.
