### 2024-11-23 #7521 @danielskatz
Thanks @jsdodge - We're having some internal discussion about which JOSS track this should be in, and once we decide, the appropriate track editor will start working on finding an editor.
