### 2024-11-29 #7521 @banesullivan
Hey howdy JOSS team, I'd like to chime in from PyOpenSci that THzTools has been accepted through PyOpenSci's review process and [I believe that helps streamline the review process for JOSS](https://www.pyopensci.org/software-peer-review/how-to/author-guide.html#journal-of-open-source-software-joss-submission).

PyOpenSci THzTools review: https://github.com/pyOpenSci/software-submission/issues/209
