### 2024-11-29 #7521 @danielskatz
@banesullivan - thanks, I missed this initially. Yes, in this case we can make this much simpler.  @crvernon or I can do the work ourselves without needed another editor.
