### 2024-12-02 #7542 @editorialbot
[REVIEW]: THzTools: data analysis software for terahertz time-domain spectroscopy
**Submitting author:** @jsdodge (<a href="http://orcid.org/0000-0003-3219-7314">J. Steven Dodge</a>)
**Repository:** https://github.com/dodge-research-group/thztools
**Branch with paper.md** (empty if default branch): main
**Version:** 0.5.5
