### 2024-12-02 #7542 @crvernon
The above review conducted under as a PyOpenSci review:

https://github.com/pyOpenSci/software-submission/issues/209
