### 2024-12-02 #7542 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
