### 2024-12-02 #7542 @crvernon
🥳 Congratulations on your new publication @jsdodge! This was published as a part of the pyOpenSci review listed in this thread.
